requirements.txt
# Core PyTorch
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Signal Processing
numpy>=1.21.0
scipy>=1.7.0

# Data Processing
pandas>=1.3.0
h5py>=3.6.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Experiment Tracking & Logging
tensorboard>=2.8.0
tqdm>=4.62.0

# Configuration Management
pyyaml>=6.0

# Machine Learning & Evaluation
scikit-learn>=1.0.0

# Wavelet Transform (required for DWT baseline)
PyWavelets>=1.1.1

# Optional: Real PTB-ECG data
wfdb>=3.4.0

# Optional: Advanced signal processing
# librosa>=0.9.0

# Testing
pytest>=6.2.0
